Limit support for x86 binaries on Windows
The binaries of x86 work perfectly in Windows of 64 bits, later it is
good to recompile all the libraries and improve the support, but now
it is not strictly necessary.

If you have a recompiled version of the binaries for 64 bits please
open a pull request or let me know in some way.

PD: GNU/Linux has x86 and x64 binaries currently working.